requirements.txt
pandas
requests
python-dotenv
matplotlib
jupyter
numpy